requirements.txt
anthropic>=0.40.0
supabase>=2.0.0
mcp>=0.9.0
fastapi>=0.110.0
uvicorn>=0.27.0
pydantic>=2.0.0
python-dotenv>=1.0.0
httpx>=0.27.0
pyyaml>=6.0.0
